Refactor HitSplat and Projectile to not be used directly (#38)
* - Refactor the combat system to use the `PerformAttack` method and streamline attack logic.
- Refactor `HitSplat` to use `HitSplatBuilder`
- Refactor `Projectile` to use `ProjectileBuilder`

* Refactor imports, remove unused references across multiple scripts, and update project references for consistency.

* Refactor NPC attack logic to use `PerformAttack` and streamline combat with builder utilities

* Refactor combat spells, NPCs, and SpellBookTab to utilize `ProjectileBuilder` and streamline attack logic

* Cleanup unused imports in TzHaar NPC scripts for consistency and maintainability.

* Refactor NPC attack scripts to use `ProjectileBuilder`, streamline combat logic, and clean up unused methods and imports.

* Update user dictionary to include new words: k'ril, tsutsaroth, waterfiend, and yarrrrrrr.

* refactor: Refactor NPC scripts and combat spells to use `ProjectileBuilder`, streamline attack logic, and clean up unused imports and methods for maintainability.

* merge: main

* fix: Streamline combat logic by refactoring NPC and spell attack scripts, removing `IncomingAttack`, consolidating `PerformAttack`, and enhancing damage handling.

+++ Hagalaz.Game.Scripts/Equipment/Barrows/AkrisaeEquipment.rseq.cs
@@ -114,15 +114,15 @@ namespace Hagalaz.Game.Scripts.Equipment.Barrows
         /// </summary>
         /// <param name="item">The current.</param>
         /// <returns></returns>
-        public override short GetDegradedItemID(IItem item)
+        public override int GetDegradedItemID(IItem item)
         {
             if (item.Id == _akrisaeIDs[0] || item.Id == _akrisaeIDs[5]
                                          || item.Id == _akrisaeIDs[10] || item.Id == _akrisaeIDs[15])
             {
-                return (short)(item.Id + 2); // Degrade to 100 barrow item
+                return item.Id + 2; // Degrade to 100 barrow item
             }
 
-            return (short)(item.Id + 1);
+            return item.Id + 1;
         }
     }
 }
